tiling

+++ rofi/install-rofi-theme.sh
@@ -9,8 +9,13 @@ if [[ $(grep -c debian /etc/os-release) != 0 ]]; then
 fi
 # install rofi theme
 git clone --depth=1 https://github.com/adi1090x/rofi.git
-cd rofi
-chmod +x setup.sh
-./setup.sh
-cd ..
-rm -rf rofi
+theme() {
+	if [[ -d ./rofi ]]; then
+		cd rofi || return
+		chmod +x setup.sh
+		./setup.sh
+		cd ..
+		rm -rf rofi
+	fi
+}
+theme
